+++ lib/services/service/event_service.dart
@@ -396,7 +396,7 @@ class EventService {
       {required BuildContext context,
       required String paymentId,
       required String eventId,
-      required int cost,
+      required double cost,
       required String dayId,
       required int person,
       required String date}) async {
