+++ lib/services/view/widgets/reservation_details_sheet.dart
@@ -574,7 +574,7 @@ void showReservationDetailsSheet({
                                                 numOfMale: maleGuestNum,
                                                 numOfFemale: femaleGuestNum,
                                                 cost: (hospitality.price *
-                                                    guestNum));
+                                                    guestNum).toDouble());
 
                                         if (isSuccess) {
                                           // Get.to(() => GeneralCheckOutScreen(
